+++ src/components/FormComponents/SignupForm/SignupForm.tsx
@@ -74,7 +74,7 @@ const SignupForm = () => {
       <Text textAlign="center" fontSize={14} color={colors.black_30}>
         Already have an account?
         <Link
-          to="/login"
+          to="/"
           style={{
             color: colors.blue_100,
             marginLeft: "5px",